+++ src/projetarm_v2/simulator/core/Cpu.java
@@ -155,7 +155,7 @@ public class Cpu {
 		}
 
 		public void hook(Unicorn u, long address, int size, Object user_data) {
-			this.cpu.pc.setValue(address);
+			this.cpu.pc.setValue((int)address);
 
 			//System.out.format(">>> Instruction @ 0x%x is being executed\n", this.cpu.pc.getValue());
